Use the URL list variables only when the user is authenticated.
The URL list variables were always passed to the template engine while
they were only defined when the user is authenticated. This patch fixes
that by using a `ctx' dictionary which is filled with common values and
with the URL list only when it is used (if the user is authenticated).