More GeoIP compatibility fixes. Create a GeoIP.GeoIP function.
The GeoIP is not exactly the same with Maxmind's library and pygeoip.
While Maxmind uses GeoIP.open(filename, flags), pygeoip uses
pygeoip.GeoIP(filename). This patch defines a GeoIP.GeoIP function which
has the same behavior as the pygeoip class constructor in order to have
enough compatibility with existant code.