More refactoring of the URL shortening code, create add_link_to_country.
This patch adds a add_link_to_country function to the views module in
order to refactor the URL shortening code. This function increments the
number of links of a country in the DB and returns the country object.
Its code was copy-pasted two times in the views, hence the refactoring.